Add option to match by relative refdes (#1)
Matching on relative refdes allows loading a sub-layout into a board
where they do not share the same schematic, for example if the
sub-layout is from a layout generator that does not have tstamps.

This is now the default (consistent with netlist import behavior), but
the previous by-tstamp is available as an option.

In replicate, changes internal indexing to by-refdes since the path is
not always available and the UUID doesn't seem available through the
API.

Also adds close-on-esc.